refactoring / consolidating code / extending functionality
