+++ DataTransformation/Core/IConnectionManager.cs
@@ -5,5 +5,6 @@ namespace Betlln.Data.Integration.Core
     public interface IConnectionManager
     {
         IDisposable GetConnection();
+        Type GetDataAdapterType();
     }
 }
